% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster.R
\name{clip_raster}
\alias{clip_raster}
\title{Clip raster}
\usage{
clip_raster(raster, shape)
}
\description{
Clips a raster to a shapefile
A user-written function from Carsten Neumann:
https://stat.ethz.ch/pipermail/r-sig-geo/2013-July/018912.html
}
