% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flowdir.R
\name{convert_flowdir}
\alias{convert_flowdir}
\title{Convert flow direction convention}
\usage{
convert_flowdir(flowdir, conv1, conv2)
}
\arguments{
\item{flowdir}{flow direction raster}

\item{conv1}{name of input flow direction convention}

\item{conv2}{name of desired output flow direction convention}
}
\description{
Converts flow direction convention among three different conventions
}
