% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_coords.R
\name{get_raster_coords}
\alias{get_raster_coords}
\title{Get raster coordinates}
\usage{
get_raster_coords(r, saveloc = FALSE)
}
\arguments{
\item{r}{raster}

\item{saveloc}{location to (optionally) save raster pixel coordinates}
}
\description{
Gets coordinates of a raster and optionally saves them as csv files
}
\examples{
basinmask <- raster("/Volumes/HD4/SWOTDA/Data/IRB/irb_mask.tif")
saveloc <- "/Volumes/HD4/SWOTDA/Data/IRB/"
domain_points <- get_domain_points(basinmask, saveloc)
}
