% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fract.R
\name{make_fract}
\alias{make_fract}
\title{Make fraction file}
\usage{
make_fract(dem, bb, saveloc, basename, target_res = 1/16)
}
\arguments{
\item{dem}{fine resolution DEM at least as large as the study area}

\item{bb}{basin boundary shapefile defining the study area}

\item{saveloc}{directory to save outputs}

\item{basename}{basename for the outputs}

\item{target_res}{target (coarse) resolution}
}
\value{
Returns the following rasters, in ASCII grid format:
Fraction File at VIC model resolution
Watershed mask at DEM resolution
Watershed mask at VIC model resolution
DEM clipped to basin boundaries
}
\description{
Creates a fraction file for the UW routing model
An alternative function is provided for creating a fraction file of all 1's
}
\details{
All coordinate systems should be geographic. I've been using WGS84.
If the pixels are not square (in geographic coordinates), this becomes a pain.
}
\examples{
dem <- raster("/Volumes/HD4/SWOTDA/Data/IRB/irb_dem.tif")
bb <- readOGR("/Volumes/HD4/SWOTDA/Data/IRB/irb_bb_1_16/bb.shp")
saveloc <- "/Volumes/HD3/SWOTDA/FDT/v12032019/Routing_Inputs"
basename <- "irb"
fract <- make_fract(dem, bb, saveloc, basename, fineres = 1/16, coarseres = 1/16)
}
