% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fract_trivial.R
\name{make_fract_trivial}
\alias{make_fract_trivial}
\title{Make fract (trivial)}
\usage{
make_fract_trivial(basinmask, saveloc, basename)
}
\arguments{
\item{basinmask}{at the modeling resolution}

\item{saveloc}{save location}

\item{basename}{basename}
}
\description{
Makes a trivial fraction file for the UW routing model, where all the fraction values are 1
}
