% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stnloc.R
\name{make_stnloc}
\alias{make_stnloc}
\title{Make stnloc}
\usage{
make_stnloc(r, gage_coords, basename, saveloc)
}
\arguments{
\item{r}{raster with the same dimensions and coordinate system as the flow
direction file (could just use flow direction direction file)}

\item{gage_coords}{text file containing lon/lat data for gage coordinaates}

\item{basename}{base name for the save file}

\item{saveloc}{location to save the station location file}
}
\description{
Makes a station location file for the UW routing model
}
\details{
Creates a station location file and saves it to the specified directory. 
Adapted from an earlier Matlab version with the same name
}
\examples{
r = raster("/Volumes/HD3/SWOTDA/Calibration/TuoSub/fdir.tif")
gage_coords = "/Volumes/HD3/SWOTDA/Calibration/TuoSub/gage_coords.txt"
saveloc <- "/Volumes/HD3/SWOTDA/Calibration/TuoSub"
basename <- "tuosub"
make_stnloc(r, gage_coords, basename, saveloc)
}
