% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_route.R
\name{report_info}
\alias{report_info}
\title{Get the mime types of the possible outputs for a report}
\usage{
report_info(file)
}
\arguments{
\item{file}{The path to the report}
}
\value{
A list with the formats, mime types, and file extensions of output,
acceptable parameters (a named list with names corresponding to the parameter
name and value corresponding to default value), and the title of the
document. For quarto documents the default values of parameters are omitted.
}
\description{
Get the mime types of the possible outputs for a report
}
\keyword{internal}
