% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routr-package.R
\docType{package}
\name{routr-package}
\alias{routr}
\alias{routr-package}
\title{routr: A Simple Router for HTTP and WebSocket Requests}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

In order to make sure that web request ends up in the correct 
    handler function a router is often used. 'routr' is a package implementing a
    simple but powerful routing functionality for R based servers. It is a fully
    functional 'fiery' plugin, but can also be used with other 'httpuv' based
    servers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/routr#routr}
  \item Report bugs at \url{https://github.com/thomasp85/routr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (0000-0002-5147-4711)

}
\keyword{internal}
