% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordlist2shape.R
\name{coordlist2shape}
\alias{coordlist2shape}
\title{coordlist2shape}
\usage{
coordlist2shape(gages, outname, r)
}
\arguments{
\item{outname}{savename for the output shapefile}

\item{r}{raster or vector object with the desired output coordinate system}

\item{coordlist}{data frame containing lat and lon values for each station}
}
\description{
Make shapefile from list of coordinates
}
