% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_route.R
\name{report_info}
\alias{report_info}
\title{Get the mime types of the possible outputs for a report}
\usage{
report_info(file)
}
\arguments{
\item{file}{The path to the report}
}
\value{
A list with mime types of output and acceptable parameters
}
\description{
Get the mime types of the possible outputs for a report
}
\keyword{internal}
